+++ Common/src/fem/meson.build
@@ -1,0 +1,8 @@
+common_src += files(['geometry_structure_fem_part.cpp',
+                     'fem_geometry_structure.cpp',
+                     'fem_integration_rules.cpp',
+                     'fem_work_estimate_metis.cpp',
+                     'fem_standard_element.cpp',
+                     'fem_wall_distance.cpp',
+                     'fem_gauss_jacobi_quadrature.cpp',
+                     'fem_cgns_elements.cpp'])
