+++ Common/src/meson.build
@@ -1,21 +1,13 @@
 common_include = include_directories('./')
-common_src =files(['geometry_structure_fem_part.cpp',
-           'graph_coloring_structure.cpp',
-           'fem_geometry_structure.cpp',
-           'fem_integration_rules.cpp',
+common_src =files(['graph_coloring_structure.cpp',
            'CConfig.cpp',
            'blas_structure.cpp',
            'ad_structure.cpp',
            'grid_movement_structure.cpp',
-           'fem_work_estimate_metis.cpp',
-           'fem_standard_element.cpp',
-           'fem_wall_distance.cpp',
            'grid_adaptation_structure.cpp',
-           'fem_gauss_jacobi_quadrature.cpp',
            'wall_model.cpp',
            'adt_structure.cpp',
            'mpi_structure.cpp',
-           'fem_cgns_elements.cpp',
            'CMultiGridQueue.cpp'])
 
 subdir('linear_algebra')
@@ -26,6 +18,7 @@ subdir('geometry/dual_grid')
 subdir('geometry/primal_grid')
 subdir('geometry/meshreader')
 subdir('interface_interpolation')
+subdir('fem')
 
 if get_option('enable-normal')
   common = static_library('SU2Common',
