+++ SU2_CFD/include/SU2_CFD.hpp
@@ -38,7 +38,7 @@
 #include "drivers/CDiscAdjMultizoneDriver.hpp"
 #include "drivers/CDummyDriver.hpp"
 #include "output/COutput.hpp"
-#include "../../Common/include/fem_geometry_structure.hpp"
+#include "../../Common/include/fem/fem_geometry_structure.hpp"
 #include "../../Common/include/geometry/CGeometry.hpp"
 #include "../../Common/include/grid_movement_structure.hpp"
 #include "../../Common/include/CConfig.hpp"
