+++ SU2_CFD/include/definition_structure.hpp
@@ -30,7 +30,7 @@
 
 #include "../../Common/include/mpi_structure.hpp"
 
-#include "../../Common/include/fem_geometry_structure.hpp"
+#include "../../Common/include/fem/fem_geometry_structure.hpp"
 #include "../../Common/include/geometry/CGeometry.hpp"
 #include "../../Common/include/CConfig.hpp"
