+++ SU2_CFD/src/output/filewriter/CSurfaceFEMDataSorter.cpp
@@ -26,7 +26,7 @@
  */
 
 #include "../../../include/output/filewriter/CSurfaceFEMDataSorter.hpp"
-#include "../../../../Common/include/fem_geometry_structure.hpp"
+#include "../../../../Common/include/fem/fem_geometry_structure.hpp"
 #include <numeric>
 
 CSurfaceFEMDataSorter::CSurfaceFEMDataSorter(CConfig *config, CGeometry *geometry, CFEMDataSorter* valVolumeSorter) :
@@ -82,7 +82,7 @@ void CSurfaceFEMDataSorter::SortOutputData() {
   }
 
   const int VARS_PER_POINT = GlobalField_Counter;
-  
+
   const unsigned long nElemLine = GetnElem(LINE);
   const unsigned long nElemTria = GetnElem(TRIANGLE);
   const unsigned long nElemQuad = GetnElem(QUADRILATERAL);
@@ -220,9 +220,9 @@ void CSurfaceFEMDataSorter::SortOutputData() {
   /* Allocate the memory for Parallel_Surf_Data. */
   nPoints = globalSurfaceDOFIDs.size();
 
-  
+
     delete [] passiveDoubleBuffer;
-  
+
 
   passiveDoubleBuffer = new passivedouble[nPoints*VARS_PER_POINT];
 
@@ -247,7 +247,7 @@ void CSurfaceFEMDataSorter::SortOutputData() {
 
   /* Determine the offset for my surface DOFs. */
   unsigned long offsetSurfaceDOFs = 0;
-  
+
   nSurfaceDOFsRanks.resize(size, 0);
 #ifdef HAVE_MPI
 
@@ -340,9 +340,9 @@ void CSurfaceFEMDataSorter::SortConnectivity(CConfig *config, CGeometry *geometr
       markerList.push_back(config->GetMarker_All_TagBound(iMarker));
     }
   }
-  
+
   /*--- Call the sort connectivity routine ---*/
-  
+
   SortConnectivity(config, geometry, markerList);
 
 }
@@ -350,7 +350,7 @@ void CSurfaceFEMDataSorter::SortConnectivity(CConfig *config, CGeometry *geometr
 void CSurfaceFEMDataSorter::SortConnectivity(CConfig *config, CGeometry *geometry, const vector<string> &markerList) {
 
   nElemPerType.fill(0);
-  
+
   SortSurfaceConnectivity(config, geometry, LINE         , markerList);
   SortSurfaceConnectivity(config, geometry, TRIANGLE     , markerList);
   SortSurfaceConnectivity(config, geometry, QUADRILATERAL, markerList);
@@ -458,8 +458,8 @@ void CSurfaceFEMDataSorter::SortSurfaceConnectivity(CConfig *config, CGeometry *
         }
       }
     }
-    
-    nElemPerType[TypeMap.at(Elem_Type)] = nSubElem_Local;    
+
+    nElemPerType[TypeMap.at(Elem_Type)] = nSubElem_Local;
 
     /*--- Store the particular global element count in the class data,
           and set the class data pointer to the connectivity array. ---*/
